Simplify color file namess, add comments, add correct alpha values to button tokens
